hierarchical classification task

+++ Dataset.py
@@ -150,7 +150,8 @@ class Dataset(Dataset):
                 frame = feature_json['frames'][index]
                 if self.empty_frame and frame['frame_id'] > len(features) * (video_fps / self.sample_fps):
                     if self.empty_frame == 'zero':
-                        features.append(np.zeros((get_points_num(is_coco=self.is_coco, body_part=self.body_part))))
+                        features.append(
+                            np.zeros((get_points_num(is_coco=self.is_coco, body_part=self.body_part))).astype(float))
                     elif self.empty_frame == 'same':
                         features.append(features[-1])
                 else:
